% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{Synchronise Project with IDHS}
\usage{
sync(from, to)
}
\arguments{
\item{from}{the path to the folder that contains the source material to sync}

\item{to}{the path to the folder where you want to sync to take place}
}
\value{
returns nothing
}
\description{
Working within the IDHS necessitates moving between two computers where one
does not have access to git. For this reason, is it useful to automate a
syncronisation process when regularly having to move between the two. Please
use with EXTREME caution, as this function overwrites and deletes files as
specified
}
\details{
MD5 hashes are used to check for file changes.

Hidden files are ignored, which is useful as we don't want to break the .git
folder anyway. Please be ultra careful about taking out any data. This is
for code sync only.
}
